Heist.Splices

Documentation

ifISplice :: Monad m => Bool -> Splice m

Run the splice contents if given condition is True, make splice disappear if not.

ifCSplice :: Monad m => (t -> Bool) -> RuntimeSplice m t -> Splice m

Function for constructing if splices that use a runtime predicate function to determine whether the node's children should be rendered.

module Heist.Splices.Apply

module Heist.Splices.Bind

module Heist.Splices.Cache

module Heist.Splices.Html

module Heist.Splices.Ignore

module Heist.Splices.Markdown