Heist.Splices.Apply

Documentation

applyTag :: Text

Default name for the apply splice.

applyAttr :: Text

Default attribute name for the apply tag.

rawApply :: Monad n => Text -> [Node] -> Maybe FilePath -> TPath -> [Node] -> Splice n

applyNodes :: MonadIO n => Template -> Text -> Splice n

Applies a template as if the supplied nodes were the children of the apply tag.

applyImpl :: MonadIO n => Splice n

Implementation of the apply splice.

deprecatedContentCheck :: Monad m => Splice m

This splice crashes with an error message. Its purpose is to provide a load-time warning to anyone still using the old content tag in their templates. In Heist 0.10, tho content tag was replaced by two separate apply-content and bind-content tags used by the apply and bind splices respectively.