Heist.Splices.Markdown

Description

The "markdown" splice formats markdown content as HTML and inserts it into the document.

If the file attribute is present the contents of the tag is ignored and the file specified is converted to HTML.

Otherwise the non-markup children of the tag are processed as markdown and converted to HTML.

This splice requires that the "pandoc" executable is in your path.

Documentation

data PandocMissingException

Constructors

PandocMissingException

data MarkdownException

Constructors

MarkdownException ByteString

data NoMarkdownFileException

Constructors

NoMarkdownFileException

markdownTag :: Text

Default name for the markdown splice.

markdownSplice :: MonadIO m => Splice m

Implementation of the markdown splice.

pandoc :: FilePath -> FilePath -> FilePath -> IO ByteString

pandocBS :: FilePath -> ByteString -> IO ByteString

readProcessWithExitCode'

Arguments

:: FilePath
command to run

-> [String]
any arguments

-> ByteString
standard input

-> IO (ExitCode, ByteString, ByteString)
exitcode, stdout, stderr