Text.Templating.Heist.Splices.Apply

Documentation

applyTag :: Text

Default name for the apply splice.

applyAttr :: Text

Default attribute name for the apply tag.

rawApply :: Monad m => [Node] -> TPath -> [Node] -> HeistT m Template

Raw core of apply functionality. This is abstracted for use in other places like an enhanced (from the original) bind

applyNodes :: Monad m => Template -> Text -> Splice m

Applies a template as if the supplied nodes were the children of the apply tag.

applyImpl :: Monad m => Splice m

Implementation of the apply splice.