Text.Templating.Heist.Splices.Cache

Documentation

data CacheTagState

State for storing cache tag information

mkCacheTag :: MonadIO m => IO (TemplateState m -> TemplateState m, CacheTagState)

Modifies a TemplateState to include a "cache" tag. The cache tag is not bound automatically with the other default Heist tags. This is because this function also returns CacheTagState, so the user will be able to clear it with the clearCacheTagState function.

clearCacheTagState :: CacheTagState -> IO ()

Clears the cache tag state.