Text.Templating.Heist.Splices.Static

Documentation

data StaticTagState

State for storing static tag information

bindStaticTag :: MonadIO m => TemplateState m -> IO (TemplateState m, StaticTagState)

Modifies a TemplateState to include a "static" tag. The static tag is not bound automatically with the other default Heist tags. This is because this function also returns StaticTagState, so the user will be able to clear it with the clearStaticTagCache function.

clearStaticTagCache :: StaticTagState -> IO ()

Clears the static tag state.